DATACMNS-1474 - Added support for Vavr 1.0.
Added dedicated build profile to test against Vavr 1.0. Adapted Try handling to accommodate the downgrade of CheckedFunction0 to not allow throwing Throwable anymore but only Exception. We now manually create Failure instances in case a Throwable occurs. Question remains of whether we should rather let Throwables bubble up and not be wrapped in a Try as they usually indicate a severe issue.